requirements.txt
PyYAML~=6.0.2
# ruamel.yaml~=0.18.10
# ruamel.yaml.clib~=0.2.8
